IAO update definitions dependent and independent variables IE about realizable -> directive controlled to remove in definition of independent variable
example of usage: For example concerning nutrition, the independent variable of your daily vitamin C intake (how much should I take) can determine the dependent variable of your life span (what is the result or observation as a result of manipulating the 'independent variable').
http://en.wikipedia.org/wiki/Dependent_and_independent_variables

example of usage. 
Controlled variable: The gender of subjects can be a controlled variable to avoid the effects of gender on the results of an investigation when it is not an independent variable.